+++ include/mainwindow.h
@@ -412,6 +412,7 @@ private:
     void updateWindowTitle();
 
     void initWindow();
+    void initLogStatusBar();
     void initCustomUI();
     void initExtraSignals();
     void initEditor();