+++ include/project.h
@@ -215,7 +215,7 @@ public:
 
     static QString getScriptFileExtension(bool usePoryScript);
     QString getScriptDefaultString(bool usePoryScript, QString mapName) const;
-    QStringList getEventScriptsFilePaths() const;
+    QStringList getEventScriptsFilepaths() const;
     void insertGlobalScriptLabels(QStringList &scriptLabels) const;
 
     QString getDefaultPrimaryTilesetLabel() const;